+++ react/__mocks__/vtex.dreamstore-icons.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react'
+
+export class IconFilter extends Component {
+  render() {
+    return <div>IconFilter</div>
+  }
+}
+
+export class IconCaret extends Component {
+  render() {
+    return <div>IconCaret</div>
+  }
+}