+++ react/__mocks__/vtex.render-runtime.js
@@ -1,0 +1,3 @@
+export function withRuntimeContext(Comp) {
+  return Comp
+}