+++ react/__tests__/components/SearchFooter.test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
 import { mount } from 'enzyme'
 import React from 'react'
-import { IconCaretLeft, IconCaretRight } from 'vtex.styleguide'
+import { IconCaret } from 'vtex.dreamstore-icons'
 
 import SearchFooter from '../../components/SearchFooter'
 
@@ -11,6 +11,7 @@ describe('<SearchFooter /> component', () => {
     recordsFiltered: 20,
     page: 1,
     maxItemsPerPage: 5,
+    runtime: { navigate: () => {} },
     getLinkProps: jest.fn(),
   }
 
@@ -28,29 +29,8 @@ describe('<SearchFooter /> component', () => {
     expect(renderComponent()).toBeDefined()
   })
 
-  it('should not render an IconCaretLeft if is the first page and render an IconCaretRight', () => {
+  it('should be render an IconCaret', () => {
     const component = renderComponent()
-    expect(component.find(IconCaretLeft).exists()).toBe(false)
-    expect(component.find(IconCaretRight).exists()).toBe(true)
-  })
-
-  it('should render an IconCaretLeft if is the last page and not render an IconCaretRight', () => {
-    const component = renderComponent({
-      page: 2,
-      recordsFiltered: 2,
-      maxItemsPerPage: 1,
-    })
-    expect(component.find(IconCaretLeft).exists()).toBe(true)
-    expect(component.find(IconCaretRight).exists()).toBe(false)
-  })
-
-  it('should render both IconCaretLeft and IconCaretRight if is neither the first nor the last page', () => {
-    const component = renderComponent({
-      page: 2,
-      recordsFiltered: 3,
-      maxItemsPerPage: 1,
-    })
-    expect(component.find(IconCaretLeft).exists()).toBe(true)
-    expect(component.find(IconCaretRight).exists()).toBe(true)
+    expect(component.find(IconCaret).exists()).toBe(true)
   })
 })